Résolution de plein de bugs. L'attaque 3D marche !!!

+++ src/base/Base.java
@@ -33,10 +33,10 @@ public class Base {
 	public Base( ReadableColor color,GameEngine gameEngine) {
 		this.gameEngine=gameEngine;
 		center=null;
-		int centerx=(int)(gameEngine.getWidth()/2);
-		int centery=(int)(gameEngine.getHeight()+(gameEngine.getWindowHeight()-gameEngine.getHeight())/2);
+		int centerx=gameEngine.getWidth()/2;
+		int centery=gameEngine.getHeight()+(gameEngine.getWindowHeight()-gameEngine.getHeight())/2;
 		this.center=new Point(centerx,centery);
-		int sizex=(int)(gameEngine.getWidth()/3);
+		int sizex=gameEngine.getWidth()/3;
 		int sizey=gameEngine.getWindowHeight()-gameEngine.getHeight();
 		this.size=new Point(sizex,sizey);
 		view = new BaseView(center,size,color);
